update ethereum SyncAccount
